Consolidate bash-tool-preference into bash-tool-guard
Both extensions steer the model away from using bash for file
operations — the preference nudges upstream before the choice is made,
the guard steers downstream after. They share a single user-facing
toggle (the bash-tool-guard resource) and a single lifecycle hook
(session_start).

Merge the standalone bash-tool-preference extension into
bash-tool-guard. The behavior is unchanged:

- TOOL_PREFERENCES_BLOCK and BASH_TOOL_DESCRIPTION constants moved into
  bash-tool-guard.ts
- toolDescriptionOverride and applyDescriptionOverride helpers
  exported from the same module
- The default extension function now registers the system prompt
  block AND mutates the bash tool description on session_start
- The standalone files are deleted
- cli.ts no longer imports the separate extension
- 19 preference tests merged into bash-tool-guard.test.ts (now 130
  tests total in that file)

Single file, single toggle, single extension lifecycle.
